[IMP]sell.py及sell_test.py拆分，以及data/view重新整合
