+++ sell/__openerp__.py
@@ -26,9 +26,9 @@
     'description': '''gooderp销售实例，通过安装gooderp模块展示odoo的销售流程''',
     'depends': ['core', 'warehouse', 'money', 'partner_address'],
     'data': [
-            'sell_data.xml',
+            'data/sell_data.xml',
             'security/groups.xml',
-            'sell_view.xml',
+            'views/sell_view.xml',
             'report/customer_statements_view.xml',
             'report/sell_order_track_view.xml',
             'report/sell_order_detail_view.xml',
@@ -47,10 +47,10 @@
             'wizard/sell_top_ten_wizard_view.xml',
             'security/ir.model.access.csv',
             'report/report_data.xml',
-            'home_page_data.xml'
+            'data/home_page_data.xml'
             ],
     'demo': [
-             'sell_demo.xml',
+             'data/sell_demo.xml',
              ],
     'installable': True,
     'auto_install': False,
