+++ sell/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+import sell_order
+import sell_delivery
+import sell_adjust
+import money
