+++ sell/tests/__init__.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
-import test_sell
+import test_sell_order
+import test_sell_delivery
+import test_sell_adjust
+import test_pricing
 import test_common_dialog
 import test_report
 import test_ReportDocx
